% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{catch22_features}
\alias{catch22_features}
\title{Catch-22 features from time series}
\usage{
catch22_features(x)
}
\arguments{
\item{tseries}{Time series of class \code{msts}.}
}
\value{
List of Catch-22 features.
}
\description{
Get the Catch-22 features from a time series of class \code{msts}.
}
