% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-load-and-processing.R
\name{cook_raw_dataframe}
\alias{cook_raw_dataframe}
\title{Cooked dataframe from raw dataframe}
\usage{
cook_raw_dataframe(
  raw_df,
  from_date,
  to_date,
  dset_key,
  filename = NULL,
  metadata = NULL
)
}
\arguments{
\item{raw_df}{Raw dataframe.}

\item{from_date}{Initial date and time of the interval. Either a \code{POSIXct} class in the GMT time zone OR a string \code{first}.}

\item{to_date}{Final date and time of the interval. Either a \code{POSIXct} class in the GMT time zone OR a string \code{last}.}

\item{dset_key}{String indicating the key of the dataset. \code{lcl} for `Low Carbon London`.}

\item{filename}{Filename.}

\item{acorn_path}{Path to the file with the ACORN values.}
}
\value{
List with the cooked dataframe, a vector of seasonal periods, a double with the percentage of \code{NA} values, and a boolean indicating if the dataframe values are all zeros. Optional members of the list are the filename and the ACORN value.
}
\description{
Cook a raw dataframe. Cooking consists in completing missing samples with NA values, removing extra samples (those not matching the sampling period), extracting a user-defined time interval out of the raw dataframe, and checking its validity to be feature-analyzed.
}
