% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Ancillary function to extract metadata}
\usage{
extract_metadata(dfs, dset_key, filename)
}
\arguments{
\item{dfs}{A list of dataframes with metadata info.}

\item{dset_key}{Dataset key: \code{lcl}, \code{goi}, etc.}

\item{filename}{A string with the file name.}
}
\value{
A list of metadata.
}
\description{
Extract metadata from files.
}
