% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{get_bins}
\alias{get_bins}
\title{Bin factors from a date sequence}
\usage{
get_bins(t, type)
}
\arguments{
\item{t}{Date sequence.}

\item{type}{Type of the bins required: 1 hour, 4 hours, 6 hours, days, weekdays/weekends, months and Northern hemisphere meteorological seasons.}
}
\value{
Factor of bins.
}
\description{
Get the bin factors from a date sequence for all possible types.
}
