% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning-analysis.R
\name{kmeans_from_features}
\alias{kmeans_from_features}
\title{k-means from features}
\usage{
kmeans_from_features(
  feats_folder,
  ftp,
  otp = NULL,
  centers,
  iter_max = 500,
  nstart = 500
)
}
\arguments{
\item{feats_folder}{String with the path to the folder of the features file.}

\item{ftp}{Vector of features to plot.}

\item{otp}{Vector of observations to plot. \code{NULL} indicates all observations.}

\item{centers}{Number of clusters.}

\item{iter_max}{Maximum number of iterations allowed.}

\item{nstart}{Number of random sets.}
}
\value{
Object of class \code{kmeans}.
}
\description{
Compute k-means from a dataframe of features.
}
