% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning-analysis.R
\name{pca_kmeans_analysis}
\alias{pca_kmeans_analysis}
\title{PCA-k-means analysis}
\usage{
pca_kmeans_analysis(feats_folder, ftp, otp = NULL, min_var = 0.95, centers)
}
\arguments{
\item{feats_folder}{String with the path to the folder of the features file.}

\item{ftp}{Vector of features to plot.}

\item{otp}{Vector of observations to plot. \code{NULL} indicates all observations.}

\item{min_var}{Minimum variance that the reduced set of principal components must fulfil (value between 0 and 1).}

\item{centers}{Number of clusters.}
}
\value{
Object of class \code{kmeans}.
}
\description{
Compute PCA from a dataframe of features, select a reduced set of dimensions for the PCA scores (depending on the acummulated explained variance provided), then computes k-means.
}
